Add files via upload
This file, along with the President.xlsx SHOULD have all the links necessary to use in populating the DB so inform me if anything is missing. FOR USA locations there will be TWO leaders, the Governor of the State, Along with President Donald Trump's image. For viewing on mobile devices, put Mr Trump's picture ABOVE the two images, and some lorem ipsum dolor that can easily be modified. He may be welcoming people or whatever we decide to put there. Hopefully he will send traffic when he see we have claimed HE is responsible for the 7 million new travel tourism jobs in that location.